v2/requirements.txt
flask>=3.0
RPi.GPIO>=0.7; platform_machine=="armv7l" or platform_machine=="aarch64"
